nekokazu/vision-app/requirements.txt
Flask==3.1.2
flask-cors==4.0.0
google-generativeai==0.8.3
gunicorn==23.0.0
functions-framework==3.7.0
Pillow==10.2.0
